The MySQL framework
This commit adds the MySQL framework and functions.
So far it checks the username of a player when they connect, if the
player is in the MySQL database it asks for a password to login, if not
it asks to create one.
The database currently stores, ID, username, IP, admin level, VIP level,
Position data, Money, Health and SkinID.
All this data is saved on player disconnect and loaded up on player
connect.